scorebird-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
tesserocr>=2.6.0
scikit-learn>=1.3.0
requests>=2.31.0
